Add files via upload
In response to this, countries worldwide have executed various Non-Pharmaceutical Interventions (NPIs) as protective/preventive measures to avoid or minimise the infection and reduce the burden on healthcare (Fisher & Heymann, 2020; Prompetchara et al., 2020). However, still, governments and authorities are facing challenges even these restrictions are imposed. The number of cases of infection, hospitalisations and mortality rates varies depending on a government's COVID-19 response, the resilience of the health system, prevalence of co-morbidities among people in a country, economic conditions, and socio-cultural factors. These factors influence people's behaviours. "For example, R0, the reproductive number, which captures the infectiousness of the pathogen, is typically measured from the growth of the epidemic and is harder to estimate once non-pharmaceutical interventions (NPIs) are in place. Similarly, changes to R0 for evolved SARS-CoV-2 variants are difficult to ascertain given simultaneous changes to behaviour and interventions" (Baker et al., 2021).

At the time of crisis, the government and authorities face challenges to manage those crises due to change in people's behaviour. Many interacting components in society motivate behavioural changes in people that influence or affect the compliance measures (NPIs) while responding to COVID-19. Behaviour change adds difficulty for epidemiologists to capture the picture of the feedback-loop between human behaviour and infectious disease (Vespignani, 2009). Therefore, it creates challenges to governments and PHAs to measure the effectiveness of those mitigation/compliance measures and understand what motivates or demotivates the public to follow/ avoid those NPIs. Thus, people's behavioural factors are the most crucial factor in understanding while studying the effectiveness of mitigation measures. Thus, in this study, I investigate what behavioural factors play an essential role in understanding individual behaviour that determines the likelihood of engaging in preventive behaviours following those NPIs.